docs(readme): update install cmd to use lockfile
Clap just released a new beta which breaks a bunch of stuff. Not looking
forward to digging through the changes to do that update.

For now, users can run 'cargo install --locked' to ensure that the
previous beta of Clap is used, thus allowing them to continue compiling
from source.

resolve #8