feat(ahk): add support for ahk2
Some users prefer to use AutoHotKey v2, so this commit adds a check for
both komorebi.ahk and komorebi.akh2 files, and the corresponding AHK
executables, whenever commands around configuration loading are run (on
startup, when manually reloading, and when watching for changes).

If both files exist in the home directory, komorebi.ahk will be
preferred (at least until AHKv2 is out of beta).

An example of a configuration file compatible with AHKv2 by @crosstyan
has been added to the documentation.

resolve #10